+++ packages/localize/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## Unreleased -->
 
-## [0.10.0] - 2021-04-18
+## [0.10.0] - 2021-04-19
 
 ### Changed